# docs/swagger.yaml
definitions:
...
  interfaces.InputLossHigh:
    properties:
      assets:
        items:
          type: string
        type: array
      loss_type:
        type: string
      maximum_loss:
        type: number
      minimum_loss:
        type: number
      most_likely_loss:
        type: number
      threat_event:
        type: string
      threat_event_id:
        type: integer
    type: object
  interfaces.InputRiskCalculator:
    properties:
      estimate_frequency:
        type: number
      estimate_loss:
        type: number
      estimate_risk:
        type: number
      max_frequency:
        type: integer
      max_loss:
        type: integer
      max_risk:
        type: integer
      min_frequency:
        type: integer
      min_loss:
        type: integer
      min_risk:
        type: integer
      mode_frequency:
        type: integer
      mode_loss:
        type: integer
      mode_risk:
        type: integer
      threat_event:
        type: string
      threat_event_id:
        type: integer
    type: object
...
  interfaces.LossHigh:
    properties:
      assets:
        items:
          type: string
        type: array
      id:
        type: integer
      loss_type:
        type: string
      maximum_loss:
        type: number
      minimum_loss:
        type: number
      most_likely_loss:
        type: number
      threat_event:
        type: string
      threat_event_id:
        type: integer
    type: object
...
  interfaces.RiskCalculator:
    properties:
      estimate_frequency:
        type: number
      estimate_loss:
        type: number
      estimate_risk:
        type: number
      id:
        type: integer
      max_frequency:
        type: integer
      max_loss:
        type: integer
      max_risk:
        type: integer
      min_frequency:
        type: integer
      min_loss:
        type: integer
      min_risk:
        type: integer
      mode_frequency:
        type: integer
      mode_loss:
        type: integer
      mode_risk:
        type: integer
      threat_event:
        type: string
      threat_event_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a server for app.
  title: Qira
  version: "1.0"
paths:
  /api/Risk:
    get:
      consumes:
      - application/json
      description: Retrieve all Risks
      parameters:
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of All Risks
          schema:
            items:
              $ref: '#/definitions/interfaces.RiskCalculator'
            type: array
      summary: Retrieve All Risks
      tags:
      - Risk
  /api/Risk/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an Risk by its ID
      parameters:
      - description: Risk ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Details
          schema:
            $ref: '#/definitions/interfaces.RiskCalculator'
      summary: Retrieve Risk by ID
      tags:
      - Risk
...
  /api/create-Risk:
    post:
      consumes:
      - application/json
      description: Create new Risk
      parameters:
      - description: Data for create new Risk
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputRiskCalculator'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Risk Create
          schema:
            $ref: '#/definitions/interfaces.RiskCalculator'
      summary: Create Risk
      tags:
      - Risk
...
  /api/create-losshigh:
    post:
      consumes:
      - application/json
      description: Create new LossHigh
      parameters:
      - description: Data for create new LossHigh
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.InputLossHigh'
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: LossHigh Create
          schema:
            $ref: '#/definitions/interfaces.LossHigh'
      summary: Create LossHigh
      tags:
      - losshigh
...
  /api/losshigh:
    get:
      consumes:
      - application/json
      description: Retrieve all LossHighs
      parameters:
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of All LossHighs
          schema:
            items:
              $ref: '#/definitions/interfaces.LossHigh'
            type: array
      summary: Retrieve All LossHighs
      tags:
      - losshigh
  /api/losshigh/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an LossHigh by its ID
      parameters:
      - description: LossHigh ID
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <token>
        description: Auth Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: LossHigh Details
          schema:
            $ref: '#/definitions/interfaces.LossHigh'
      summary: Retrieve LossHigh by ID
      tags:
      - losshigh
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"